Create Genesis 1:1
The Torah was originally written in Sanskrit and the Torah is about the AiLaVaṃSha. The Torah was not originally written in Hebrew. Thankfully the Hebrew Torah retains the correct sequence of letters. To translate the Torah from Hebrew into Sanskrit the vowels must be reattached the Hebrew letters. According to Hebrew Torah scholars the Torah was originally written as a sequence of letters without spaces between the words and between the paragraphs; and traditionally that is the way Sanskrit is written. 

The contemporary Hebrew script is also not the original script of the Torah. Hebrew Torah scholars have proposed that the Torah was originally written in the Ancient Hebrew script. I propose that the Ancient Hebrew script is descended from Sanskrit which is the original language of the Torah. Sanskrit may be the source language from which Hebrew developed. The reason that scholars are unsure of how to pronounce the Ancient Hebrew of the Torah is because the Sanskrit vowel symbols attached to the Sanskrit letters are missing. 

In Sanskrit the vowel symbols are attached to the letters and every letter initially has a vowel sound of “short a” attached to the letter unless the letter is a vowel. When the Torah was translated from Sanskrit into Hebrew those vowel sounds connected to the Sanskrit letters were lost. That initial vowel sound can change to “long a,” “short and long i,u, and r,” and “l,” “e,” “ai,” “o,” and “au” depending on the addition of special characters connected to the contemporary Sanskrit letters. To translate the Torah from Hebrew into Sanskrit the correct Sanskrit vowels must be reattached to the Hebrew letters. For example, the word for “God” in the Holy Bible in Genesis 1:1 comes from the Hebrew word "ALHYM" in the Hebrew Torah which may have been translated from the Sanskrit words "AiLa ḥ JāMi." By reattaching the Sanskrit vowels to the Hebrew letters in the Hebrew word “ALHYM” the first Hebrew letter “A” may be the Sanskrit letter “Ai,” the Hebrew “L” may be the Sanskrit “La,” the Hebrew “H” may be the Sanskrit “ḥ,” the Hebrew “Y” which may also be a “J” may be the Sanskrit “Jā,” and the Hebrew “M” may be the Sanskrit “Mi.” If these are the correct vowels then the Sanskrit words "AiLa ḥ JāMi" may be formed. The Sanskrit word “AiLa” refers to the “AiLaVaṃSha.” “VaṃSha” means “tribe.” The “AiLaVaṃSha” are the “Lunar Dynasty” of ancient India. The Sanskrit word “JāMi” means "related." "AiLa ḥ JāMi" may mean “related” to the “AiLaVaṃSha.” For example again, the word for “Heaven” in the Holy Bible comes from the Hebrew word “SMYM” in the Hebrew Torah which may have been translated from the Sanskrit words “SoMa JāMi.” The Hebrew “S” may become the Sanskrit “So,” the Hebrew “M” may become the Sanskrit “Ma,” and the Hebrew “YM” may become the Sanskrit word “JāMi.” The Sanskrit word “SoMa” may mean the “Moon.” In Sanskrit the Torah may be describing the “Lunar Dynasty” and the “Moon.” The Sanskrit Torah may not be describing “God” creating “Heaven” and Earth. 

Genesis 1:1 word one is traditionally the first word of the Hebrew Torah spelled “BRA” ; And to translate from Hebrew into Sanskrit the Sanskrit vowels must be reattached to the Hebrew letters. There are a variety of vowels which may be correct so it is difficult to identify the correct Sanskrit vowel for any Hebrew letter. Ideally the original Sanskrit Torah must be recovered. “BRA” may become “BhaRaAi” and may be represented by the contemporary Sanskrit word “BhaRi” meaning “nourish.” The ancient word “BhaRaAi” could have a similar definition of “nourishing,” “sustaining,” or “nurturing.”

Genesis 1:1 word two is the Hebrew letters “SJT” which is traditionally connected to “BRA” forming “BRASJT” which is the traditionally accepted first word of Genesis meaning “In the Beginning” in the Holy Bible. What exactly is being formed or nourished, and who is forming what it is? The Sanskrit vowels must be reattached to “SJT” which may become the Sanskrit word “SaJāTa” meaning “countrymen.” 

Genesis 1:1 words one through two may mean “forming countrymen” or “sustaining countrymen.” Genesis 1:1 words one through two may then be written as “BhaRaAiSaJāTa.” Another treasure is formed from the conjunction of “BhaRaAi” with “SaJāTa” which is the Sanskrit word “AiSa” meaning “divine,” and possibly all together meaning “sustaining divine countrymen.” Genesis 1:1 word three is the Hebrew word “BRA” restated. 

Genesis 1:1 word four is the most important word of the Holy Bible which is traditionally “ALHYM” which may become the Sanskrit “AiLa ḥ JāMi.” The contemporary Sanskrit words “AiLa ḥ JāMi” represents the tribe of “AiLa.” The tribe of “AiLa” are said to be the progeny of the mother “Ila,” and “Ila” is said to be the progeny of “MaNu.”

“Indra” means “Lord.” When the letters that spell “Israel” are converted into Sanskrit then the Hebrew word “ISRAL” may spell the Sanskrit words “JaSuRi AiLa.” The Hebrew “I” may also be a “Y” or a “J.” “JaSuRi AiLa” means “Indra’s thunderbolt of the AiLaVaṃSha.” 

Genesis 1:1 word five in Hebrew is “AT” and may become the Sanskrit word “AiTa” followed by the silent “ḥ” in the Sanskrit Torah. “AiTa” may be similar to “ETa” in contemporary Sanskrit meaning “these.” Genesis 1:1 word six in Hebrew is “SMYM” and may become the two Sanskrit words “SoMa JāMi” meaning the “Moon” or “outer space” and “related.” Genesis 1:1 word seven in Hebrew is “VAT” which in Sanskrit may become “Va AiTa” again followed by “ḥ.” Genesis 1:1 word eight in Hebrew is “ARTz” and translating that into Sanskrit is difficult. The Hebrew “ARTz” may become the Sanskrit words “AiRa” and “DSha.” “AiRa” may mean “consisting of water or refreshment” meaning “abundant,” and the Hebrew “Tz” could be the contemporary Sanskrit word “DeSha” meaning “country.” “AiRa DSha” could mean “abundant country.” Genesis 1:1 words one through eight may thus be “BhaRaAi SaJāTa BhaRaAi AiLaḥ JāMi AiTaḥ SoMa JāMi Va AiTaḥ AiRa DeSha” meaning “nourishing countrymen nourishing the Lunar Dynasty these of the Moon and these of the abundant country.”